+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+* Core: Stop MONITOR and PubSub streams from dropping buffered handshake bytes ([#6979](https://github.com/valkey-io/valkey-glide/pull/6979))
 * Java: Fix scoped connection truncating values larger than 16 KB ([#6893](https://github.com/valkey-io/valkey-glide/issues/6893))
 * Go: Propagate pool ConnectionRequest into pool-borrowed clients so `ScopedConnection` works on pooled clients ([#6763](https://github.com/valkey-io/valkey-glide/issues/6763))
 * Core/FFI: Standalone AZ-affinity reads skip nodes that are reconnecting instead of blocking on them; accept `AllNodes` in `create_client_from_uri`'s `read_from` option ([#6721](https://github.com/valkey-io/valkey-glide/pull/6721))